Синхронизация правок 1: не перемещенные файлы

+++ xray/xrCore/_types.h
@@ -15,7 +15,7 @@ typedef	signed		__int64	s64;
 typedef	unsigned	__int64	u64;
 
 typedef float				f32;
-typedef double				f64;
+typedef long double			f64;
 
 typedef char*				pstr;
 typedef const char*			pcstr;
