+++ xray/xr_3da/xrGame/actor_statistic_mgr.h
@@ -13,7 +13,7 @@ public:
 	SStatSectionData&		GetSection				(const shared_str& key);
 	
 	void					AddPoints				(const shared_str& key, const shared_str& detail_key, const shared_str& str_value);
-	void					AddPoints				(const shared_str& key, const shared_str& detail_key, s32 cnt, s32 pts);
-	s32						GetSectionPoints		(const shared_str& key);
+	void					AddPoints				(const shared_str& key, const shared_str& detail_key, s32 cnt, float pts);
+	float					GetSectionPoints		(const shared_str& key);
 	const vStatSectionData&	GetCStorage				();
 };
